requirements.txt
aniso8601==8.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
ffmpeg-python==0.2.0
Flask==1.0.2
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
flask-mongoengine==0.9.5
Flask-RESTful==0.3.6
Flask-WTF==0.14.2
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
marshmallow==2.15.4
mongoengine==0.18.2
PyJWT==1.7.1
pymongo==3.7.1
pytz==2019.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
bcrypt==3.1.7
pyOpenSSL==19.1.0
